// src/app/data/interfaces/confPagos.ts
export interface confPagos  {
    id?: any;
    id_fraccionamiento?: any;
    descripcion?: any;
    monto?: any;
    periodo?: any;
    tipo_pago?: any;
    porcentaje_descuento?: any;
    porcentaje_penalizacion?: any;
    dias_max_descuento?: any;
    dias_max_pago?: any;
    fecha_inicial?: any;
    estatus?: any;
    created_at?: any;
    updated_at?: any;
  }
  